pyproject.toml
[project]
name = "news-aggregator"
version = "0.1.0"
requires-python = ">=3.10,<3.14"
dependencies = [
    "supabase==2.29.0",
    "python-dotenv==1.2.2",
    "anthropic>=0.40.0",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.1.0",
    "psycopg2-binary>=2.9.12",
]
